removed compiler warnings from my latest fix

+++ src/main.c
@@ -500,7 +500,7 @@ struct Configurable {
   bool resume_from_current;
   bool disable_epsv;
   bool disable_eprt;
-  int resume_from;
+  long resume_from;
   char *postfields;
   long postfieldsize;
   char *referer;
@@ -513,8 +513,8 @@ struct Configurable {
   char *iface;
   unsigned short porttouse;
   char *range;
-  int low_speed_limit;
-  int low_speed_time;
+  long low_speed_limit;
+  long low_speed_time;
   bool showerror;
   char *userpwd;
   char *proxyuserpwd;
@@ -1060,7 +1060,7 @@ static void cleanarg(char *str)
  * The string must start with a digit to be valid.
  */
 
-static int str2num(int *val, char *str)
+static int str2num(long *val, char *str)
 {
   int retcode = 0;
   if(isdigit((int)*str))
